# packages/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LangAI",
    "description": "Private dictionary representation training. Lexical knowledge is not conversational fluency.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/languages/{language_id}/training-runs": {
      "post": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Train",
        "operationId": "train_languages__language_id__training_runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "language_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Runs",
        "operationId": "runs_languages__language_id__training_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "language_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training-runs/{run_id}": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Run",
        "operationId": "run_training_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training-runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Cancel",
        "operationId": "cancel_training_runs__run_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training-runs/{run_id}/events": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Events",
        "operationId": "events_training_runs__run_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training-runs/{run_id}/stream": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Stream",
        "operationId": "stream_training_runs__run_id__stream_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/languages/{language_id}/models": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Models",
        "operationId": "models_languages__language_id__models_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "language_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Model",
        "operationId": "model_models__model_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/evaluation": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Evaluate",
        "operationId": "evaluate_models__model_id__evaluation_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/lineage": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Lineage",
        "operationId": "lineage_models__model_id__lineage_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/approve": {
      "post": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Approve",
        "operationId": "approve_models__model_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/artifact": {
      "get": {
        "tags": [
          "Training & registry"
        ],
        "summary": "Artifact",
        "operationId": "artifact_models__model_id__artifact_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hyperparameters": {
        "properties": {
          "epochs": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Epochs"
          },
          "batch_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 512.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size"
          },
          "learning_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 0.1,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Learning Rate"
          },
          "device": {
            "type": "string",
            "enum": [
              "auto",
              "cpu",
              "cuda",
              "mps"
            ],
            "title": "Device",
            "default": "auto"
          },
          "patience": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Patience",
            "default": 2
          },
          "max_length": {
            "type": "integer",
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Max Length",
            "default": 128
          },
          "weight_decay": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight Decay",
            "default": 0.01
          },
          "evaluation_frequency": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Evaluation Frequency",
            "default": 1
          },
          "margin": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Margin",
            "default": 0.4
          },
          "lora_rank": {
            "type": "integer",
            "maximum": 64.0,
            "minimum": 2.0,
            "title": "Lora Rank",
            "default": 8
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Hyperparameters"
      },
...
      "TrainCreate": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "task": {
            "type": "string",
            "enum": [
              "embeddings",
              "language-identification",
              "dictionary-adapter"
            ],
            "title": "Task",
            "default": "embeddings"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model",
            "default": "auto"
          },
          "base_model_revision": {
            "type": "string",
            "title": "Base Model Revision",
            "default": "auto"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "auto",
              "contrastive",
              "char-ngram",
              "lora"
            ],
            "title": "Strategy",
            "default": "auto"
          },
          "random_seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Random Seed",
            "default": 42
          },
          "hyperparameters": {
            "$ref": "#/components/schemas/Hyperparameters"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "TrainCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}